Updates match scouting functionality.
Makes it so that if the user sets AQ2 (Hang gears?) to true, it sets the
value of AQ3 (Cross line?) to true, as well.

Updates BurgerMinder to remove the graphic, add a legend at the top, and
changes the questions that are required to match the updates made in
Match scouting. Also changes to remove % completed, relying instead on
the list of unanswered questions added earlier. Added a "None" value to
the list of unanswered questions for when all of them have been
answered.

Changes AQ10 from a drop-down to a set of exclusive checkboxes (similar
to radio buttons). Changes the export function and other functions to
accommodate the change to the field type.

Changes exported checkbox data from [blank] and TRUE to "0" and "1".